# openapi.yaml
openapi: 3.1.0
info:
  title: AgriCRM API
  description: Farmer, FPO & Sugar Mill CRM for Theta Analytics
  version: 1.0.0
paths:
...
  /api/v1/gstin/check/:
    get:
      tags:
      - gstin
      summary: Gstin Local Check
      description: "Layer one: format, embedded PAN, state code and checksum. No network.\n\
        \n\U0001F534 Fast enough to run as the user types, and it is **not** a verification.\n\
        The response says so in `note`, and there is deliberately no field a UI\n\
        could read as \"GST-verified\"."
      operationId: gstin_local_check_api_v1_gstin_check__get
      security:
      - HTTPBearer: []
      parameters:
      - name: value
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 32
          title: Value
      - name: govt_uin
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Govt Uin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalCheckOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gstin/verifications/:
    post:
      tags:
      - gstin
      summary: Gstin Verification Create
      description: "Layer two: ask the provider whether this registration is active.\n\
        \nCached for a configurable TTL and deduplicated across concurrent callers.\n\
        `force` is what **Verify again** calls \u2014 a registration can be cancelled\n\
        without notice, and a customer that matters is worth a fresh lookup.\n\n\U0001F534\
        \ Only the GSTIN is sent to the provider. Never invoice lines, never\namounts,\
        \ never anything else from the CRM."
      operationId: gstin_verification_create_api_v1_gstin_verifications__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/gstin/verifications/{verification_id}/:
    get:
      tags:
      - gstin
      summary: Gstin Verification Get
      operationId: gstin_verification_get_api_v1_gstin_verifications__verification_id___get
      security:
      - HTTPBearer: []
      parameters:
      - name: verification_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Verification Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/invoices/{invoice_id}/gstin-check/:
    post:
      tags:
      - gstin
      summary: Invoice Gstin Check
      description: "Both layers for this invoice's buyer, plus the CRM comparison.\n\
        \n\U0001F534 The comparison reports differences and writes none. Silently\
        \ overwriting\na customer record with a provider's spelling of their name\
        \ is how a\nregistry stops being something a human curated."
      operationId: invoice_gstin_check_api_v1_invoices__invoice_id__gstin_check__post
      security:
      - HTTPBearer: []
      parameters:
      - name: invoice_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Invoice Id
      - name: force
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Force
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GstinCheckOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/invoices/{invoice_id}/gstin-check/use-verified/:
    post:
      tags:
      - gstin
      summary: Invoice Gstin Use Verified
      description: "Populate a **draft** from the verified registry identity.\n\n\U0001F534\
        \ Draft only, and only for an active registration. An issued invoice's\nbuyer\
        \ block is a snapshot of what was printed; a customer that moved office\n\
        must not silently alter a document their accounts team already holds."
      operationId: invoice_gstin_use_verified_api_v1_invoices__invoice_id__gstin_check_use_verified__post
      security:
      - HTTPBearer: []
      parameters:
      - name: invoice_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Invoice Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UseVerifiedResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/invoices/{invoice_id}/gstin-check/override/:
    post:
      tags:
      - gstin
      summary: Invoice Gstin Override
      description: "\U0001F534 Proceed despite an unavailable or mismatched verification.\n\
        \nSeparately permissioned, and it captures actor, reason and time on an\n\
        immutable row. Allowed only where the result is *unknown* or a mismatch \u2014\
        \nnever where the registry positively reports the registration cancelled,\n\
        because billing GST to a cancelled registration is not a judgement call."
      operationId: invoice_gstin_override_api_v1_invoices__invoice_id__gstin_check_override__post
      security:
      - HTTPBearer: []
      parameters:
      - name: invoice_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Invoice Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GstinOverrideRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GstinCheckOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GstinCheckOut:
      properties:
        gstin:
          anyOf:
          - type: string
          - type: 'null'
          title: Gstin
        local:
          $ref: '#/components/schemas/LocalCheckOut'
        live:
          anyOf:
          - $ref: '#/components/schemas/VerificationOut'
          - type: 'null'
        differences:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Differences
        policy:
          type: string
          title: Policy
        blocks_issue:
          type: boolean
          title: Blocks Issue
        overridden:
          type: boolean
          title: Overridden
          default: false
      type: object
      required:
      - gstin
      - local
      - live
      - differences
      - policy
      - blocks_issue
      title: GstinCheckOut
    GstinOverrideRequest:
      properties:
        reason:
          type: string
          maxLength: 1000
          minLength: 10
          title: Reason
      type: object
      required:
      - reason
      title: GstinOverrideRequest
      description: "\U0001F534 A reason is mandatory, and it is stored on an immutable\
        \ row with the\nactor and the time. An override nobody can review afterwards\
        \ is not a\ncontrol, it is a bypass."
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LocalCheckOut:
      properties:
        supplied:
          type: string
          title: Supplied
        normalised:
          anyOf:
          - type: string
          - type: 'null'
          title: Normalised
        valid:
          type: boolean
          title: Valid
        is_govt_uin:
          type: boolean
          title: Is Govt Uin
        state_code:
          anyOf:
          - type: string
          - type: 'null'
          title: State Code
        state_name:
          anyOf:
          - type: string
          - type: 'null'
          title: State Name
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        note:
          type: string
          title: Note
      type: object
      required:
      - supplied
      - normalised
      - valid
      - is_govt_uin
      - state_code
      - state_name
      - message
      - note
      title: LocalCheckOut
      description: "Layer one only.\n\n\U0001F534 There is deliberately no field here\
        \ a UI could render as \"GST-verified\".\n`valid` means well-formed; whether\
        \ the registration is active is a\ndifferent question with a different endpoint,\
        \ and conflating them is the\ndefect this module exists to prevent."
...
    UseVerifiedResult:
      properties:
        invoice_id:
          type: string
          format: uuid
          title: Invoice Id
        verification:
          $ref: '#/components/schemas/VerificationOut'
        changes:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Changes
      type: object
      required:
      - invoice_id
      - verification
      - changes
      title: UseVerifiedResult
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VerificationCreate:
      properties:
        billing_entity:
          type: string
          format: uuid
          title: Billing Entity
        gstin:
          type: string
          maxLength: 32
          minLength: 1
          title: Gstin
        govt_uin:
          type: boolean
          title: Govt Uin
          default: false
        force:
          type: boolean
          title: Force
          default: false
      type: object
      required:
      - billing_entity
      - gstin
      title: VerificationCreate
    VerificationOut:
      properties:
        id:
          type: string
          title: Id
        gstin:
          type: string
          title: Gstin
        provider:
          type: string
          title: Provider
        provider_reference:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider Reference
        status:
          type: string
          title: Status
        is_verified:
          type: boolean
          title: Is Verified
        is_unavailable:
          type: boolean
          title: Is Unavailable
        legal_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Legal Name
        trade_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Trade Name
        registration_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Registration Type
        taxpayer_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Taxpayer Status
        effective_from:
          anyOf:
          - type: string
          - type: 'null'
          title: Effective From
        cancellation_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Cancellation Date
        principal_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Principal Address
        state_code:
          anyOf:
          - type: string
          - type: 'null'
          title: State Code
        checked_at:
          type: string
          title: Checked At
        expires_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Expires At
        age_days:
          type: integer
          title: Age Days
        raw_response_sha256:
          anyOf:
          - type: string
          - type: 'null'
          title: Raw Response Sha256
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        error_detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Detail
        label:
          type: string
          title: Label
      type: object
      required:
      - id
      - gstin
      - provider
      - provider_reference
      - status
      - is_verified
      - is_unavailable
      - legal_name
      - trade_name
      - registration_type
      - taxpayer_status
      - effective_from
      - cancellation_date
      - principal_address
      - state_code
      - checked_at
      - expires_at
      - age_days
      - raw_response_sha256
      - error_code
      - error_detail
      - label
      title: VerificationOut
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer